requirements.txt
pyyaml
requests
pyarrow
fastparquet
pandas
pytest
pytest-mock